[#1835] Prototype non-merging schema
